Create TCGplayer Cart card list with prices
This script will take your current TCGplayer cart and parse it for the price and card name of each item in the cart. This is helpful for when you know your deck list but not necessarily to cost of each card. For me it was helpful in determining which cards to proxy and which to buy. 
Copy this script
go to the TCGplayer cart
press F12
paste the script in the console and press enter
right-click the red text
click "Copy String Contents"
Paste in Notepad++ (I used notepad++ use what you know/have)
In notepad++: edit -> line operations -> sort lines as integers descending

you can then use the data right in notepad++ or save it as a .csv and upload it to sheets or excel

And BOOM, you got your cart from TCGplayer into a usable list.

***Disclaimer*** I am a knuckle dragging mouth-breather, if it doesn't work I can't help you.
 I DID NOT WRITE THIS CODE 
I like to think I helped a little but a friend only known as Joe did all the hard work and I give full credit to him, Thank you Joe.